Outsourced Settings.jsx into components
